MB-25 Modified primary score to use double

+++ src/main/java/com/bhf/jmhresults/JMHResult.java
@@ -1,12 +1,11 @@
 package com.bhf.jmhresults;
 
-import java.math.BigDecimal;
 import org.json.simple.JSONObject;
 
 public class JMHResult {
 
 	public final String benchmarkName;
-	public final BigDecimal primaryScore;
+	public final double primaryScore;
 	public final JSONObject secondaryScores;
 	public final JSONObject params;
 	public double L1_dcache_loads;
@@ -24,11 +23,12 @@ public class JMHResult {
 	public double STALLS_L2_PENDING;
 	public double STALLS_LDM_PENDING;
 
-	public JMHResult(String benchmarkName, BigDecimal primaryScore, JSONObject secondaryScores, JSONObject params) {
+	public JMHResult(String benchmarkName, double primaryScore, JSONObject secondaryScores, JSONObject params) {
 		this.benchmarkName = benchmarkName;
 		this.primaryScore = primaryScore;
 		this.secondaryScores = secondaryScores;
 		this.params = params;
-
 	}
+	
+	
 }
